fix(proxy): move runtime teardown helper to internal test support
