+++ Tests/XcodeMCPProxyTestSupport/AsyncTestSupport.swift
@@ -3,8 +3,6 @@ import Foundation
 import NIO
 import NIOConcurrencyHelpers
 
-@testable import XcodeMCPProxyKit
-
 package struct AsyncTestTimeoutError: Error, CustomStringConvertible {
     package let description: String
 
@@ -422,29 +420,6 @@ package func shutdown(_ group: EventLoopGroup) async {
         group.shutdownGracefully { _ in
             continuation.resume()
         }
-    }
-}
-
-extension RuntimeCoordinator {
-    /// Test-only synchronous teardown for defer blocks; production code
-    /// awaits shutdown() directly.
-    package func shutdownAndWait() {
-        let semaphore = DispatchSemaphore(value: 0)
-        Task.detached(priority: .userInitiated) { [self] in
-            await shutdown()
-            semaphore.signal()
-        }
-        semaphore.wait()
-    }
-
-    package func drainRuntimeTasksAndWaitForTesting() {
-        let drain = runtimeTasks.drainCurrentTasks()
-        let semaphore = DispatchSemaphore(value: 0)
-        Task.detached(priority: .userInitiated) {
-            await drain.wait()
-            semaphore.signal()
-        }
-        semaphore.wait()
     }
 }
